migrate(tools): relocate the agent_tools dispatch family into the tools module (Class A #2)
Second Class A migration. The required tools module (core capability #18, tool dispatch) gets its
implementation: the posix agent_tools* family and its contract header move into src/modules/tools/
and the descriptor latches. 20/26 latched; six Class A modules remain.

Decision (self-reviewed; roundtable unavailable this cycle, under operator authorization for
self-review-only): the tool code is distributed with a name collision — src/posix/agent_tools.c
(implementations) and src/server/agent_tools.c (turn/snapshot/toolset session-state) share a filename,
and agent_tools.h declares symbols implemented in both. Resolution:

- tools owns the dispatch/implementation surface: agent_tools.c, agent_tools_dispatch.c,
  agent_tools_anchored.c + the contract agent_tools.h + the private seam agent_tools_internal.h,
  relocated from src/posix/ and src/headers/ under src/modules/tools/.
- src/server/agent_tools.c is LEFT in the server as a not-module-local implementation of the
  session-state slice of agent_tools.h — the same arrangement by which memory owns its contract while
  DB1/DB2 implement storage (slice 56). Only the posix agent_tools.c moves, so the name collision
  never arises. The two halves communicate only through public agent_tools.h; server/agent_tools.c
  does not include the posix private seam, so no new coupling.

Build: agent_tools.h leaves the global src/headers/ dir, so -Imodules/tools is added to the Make
C_FLAGS (one global flag covers all 30 includers incl src/windows/) and to the six CMake module
include lists; source/object/link references repoint posix/agent_tools* -> modules/tools/agent_tools*;
one path-qualified include in test_cmd_delegate.c is rewritten. agent_tools.h is declared a
private_header per the flat-layout convention; tests declares test_tool_output_cap.c (its subject
agent_tool_output_cap_clamp is a tools-contract inline). Docs/register/ledger/test-baseline updated;
slice-59 validation doc added; the empty-root test drops tools and gains its latched mutation coverage.

Verified: all module/descriptor checks + the descriptor test suite pass; aimee-server builds and links
the relocated objects and every agent_tools.h consumer. Full Make + CMake .253 validation to follow.
