+++ src/modules/tools/agent_tools.c
