+++ src/modules/tools/agent_tools.h
