+++ src/modules/tools/agent_tools_anchored.c
