+++ src/modules/tools/agent_tools_internal.h
