+++ src/tests/test_cmd_delegate.c
@@ -10,7 +10,7 @@
 #include <aimee/delegates/delegate_role.h>
 #include "model_registry.h"
 #include "log.h"
-#include "posix/agent_tools_internal.h"
+#include "agent_tools_internal.h"
 #include "provider_cli_adapter.h"
 #include "cJSON.h"
